Use `ash`' built-in `"linked"` feature for static linking
All the code here to define an `extern "system"` symbol for
`vkGetInstanceProcAddr` and use it inside `Entry` (via `StaticFn`) was
included in Ash release `0.35` (the minimum compatbile with `ash-molten`
already) to make it much more trivial to static-link on various
platforms, including against `MoltenVK` here.

This change makes the `lib.rs` portion of `MoltenVK` pretty much
useless as it's only forwarding the call (and `load()` should perhaps
be renamed?), should we rethink the structure of the crate a bit.